pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "osu-sync"
version = "0.1.0"
description = "osu! beatmap synchronization tool"
authors = [{name = "AkaakuHub", email = "akaaku2021media01@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "aiolimiter>=1.1.0",
    "pydantic>=2.5.0",
    "pywebview>=4.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
osu-sync = "src.main:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.hatch.build.targets.wheel]
packages = ["src"]